+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - fix the `addAnimation()` method not returning the corresponding set TrackEntry
 - fix the base renderable `flip[X/Y]` method when used/applied to the Spine renderable
 - add a `isCurrentAnimation()` that returns true if the give name corresponding to the current track current animation
+- expose the `currentTrack` property to access the corresponding current animation track entry
 
 ## 1.4.0 - 2023-09-05